requirements.txt
pypdf2
langchain==0.1.1
langchain_community==0.0.13
langchain_openai==0.0.2.post1
openai==1.8.0
python-dotenv==1.0.0
streamlit==1.30.0
faiss-cpu
tiktoken
pandas
scikit-learn==1.2.1 
rank_bm25
nltk
sentence_transformers
transformers
datasets
langchain-core 
langchain-mistralai
qdrant-client
fastembed
weaviate-client==3.*
pandas
scikit-learn
numpy
scipy
pickle-mixin
tqdm
rank_bm25
azure-search-documents==11.4.0b6
tenacity
torchvision
plotly
nbformat
nltk